fix(header): fix header insert location
Previously, if a header did not exist, it was inserted one line too
early (potentially breaking a previous "block" of hosts in the file).
It also added an extra blank line after the header section.

The header line is now added in the correct position, with a
preceeding blank line.

Additionally, a fixtures file for archlinux and simple use case was
added too.